refactor(utils): add getDiffInCalendarDays as ESM
